Move print_msg() to the 'osc.output' module
