+++ osc/output/__init__.py
@@ -1,5 +1,6 @@
 from .key_value_table import KeyValueTable
 from .input import get_user_input
+from .output import print_msg
 from .tty import colorize
 from .widechar import wc_ljust
 from .widechar import wc_width
